Reorganize internal documents
