+++ docs/internal/investigations/assets/concept.png
