+++ docs/internal/investigations/runtime-director-poc/component/Makefile
@@ -1,5 +1,5 @@
 APP_NAME = runtime-director
-APP_PATH = docs/internal/investigations/runtime-director-dapr/component
+APP_PATH = docs/internal/investigations/runtime-director-poc/component
 ENTRYPOINT = cmd/main.go
 BUILDPACK = eu.gcr.io/kyma-project/test-infra/buildpack-golang-toolbox:v20200423-1d9d6590
 SCRIPTS_DIR = $(realpath $(shell pwd)/../../../../../)/scripts
