+++ docs/internal/investigations/runtime-director-poc/component/go.mod
@@ -1,4 +1,4 @@
-module github.com/kyma-project/control-plane/docs/internal/investigations/runtime-director-dapr/component
+module github.com/kyma-project/control-plane/docs/internal/investigations/runtime-director-poc/component
 
 go 1.14
